+++ src/main/java/minigame/cryptocurrency/CryptoCurrency.java
@@ -1,4 +1,4 @@
-package minigame.CryptoCurrency;
+package minigame.cryptocurrency;
 
 import player.PlayerProfile;
